Fix receiver errors and warnings

+++ lib/f_core/radio/frame_handlers/c_frequency_change_handler.cpp
@@ -15,7 +15,7 @@ void CFrequencyChangeHandler::HandleFrame(const LaunchLoraFrame& frame) {
 
     LOG_INF("Changing frequency to %f Hz", static_cast<double>(newFrequencyMhz));
     if (!lora.SetFrequency(newFrequencyMhz)) {
-        LOG_ERR("Failed to set new frequency %f Hz", newFrequencyMhz);
+        LOG_ERR("Failed to set new frequency %f Hz", static_cast<double>(newFrequencyMhz));
         return;
     }
